+++ deprecated_package/lib/eatshots_video_player_method_channel.dart
@@ -1,6 +1,0 @@
-export 'package:video_view_player/video_view_player_method_channel.dart';
-
-import 'package:video_view_player/video_view_player_method_channel.dart';
-
-@Deprecated('Use MethodChannelVideoViewPlayer instead')
-typedef MethodChannelEatshotsVideoPlayer = MethodChannelVideoViewPlayer;