+++ deprecated_package/lib/eatshots_video_player_platform_interface.dart
@@ -1,6 +1,0 @@
-export 'package:video_view_player/video_view_player_platform_interface.dart';
-
-import 'package:video_view_player/video_view_player_platform_interface.dart';
-
-@Deprecated('Use VideoViewPlayerPlatform instead')
-typedef EatshotsVideoPlayerPlatform = VideoViewPlayerPlatform;