+++ deprecated_package/lib/eatshots_video_player_pool.dart
@@ -1,6 +1,0 @@
-export 'package:video_view_player/video_view_player_pool.dart';
-
-import 'package:video_view_player/video_view_player_pool.dart';
-
-@Deprecated('Use VideoViewPlayerPoolManager instead')
-typedef EatshotsVideoPlayerPoolManager = VideoViewPlayerPoolManager;