+++ deprecated_package/lib/eatshots_video_player_widget.dart
@@ -1,6 +1,0 @@
-export 'package:video_view_player/video_view_player_widget.dart';
-
-import 'package:video_view_player/video_view_player_widget.dart';
-
-@Deprecated('Use VideoViewPlayer instead')
-typedef EatshotsVideoPlayer = VideoViewPlayer;